外层循环距离问题之: 视觉算法距离值改为整型;

+++ SMG_NothingIsAll/CortexAlgorithms/Vision/AIVisionAlgsModel.h
@@ -34,7 +34,7 @@
 @property (assign,nonatomic) CGFloat direction;
 
 //distance
-@property (assign,nonatomic) CGFloat distance;
+@property (assign,nonatomic) NSInteger distance;
 
 //border
 @property (assign,nonatomic) CGFloat border;
